Test.QuickCheck.ContractModel.Internal.Spec

Documentation

data ModelState state

The ModelState models the state of the blockchain. It contains,

the contract-specific state (contractState)

the current slot (currentSlot)

the wallet balances (balances)

the amount that has been minted (minted)

Constructors

ModelState

Fields

_currentSlot :: SlotNo

_balanceChanges :: Map (AddressInEra Era) SymValue

_minted :: SymValue

_symbolics :: SymCollectionIndex

_assertions :: [(String, Bool)]

_assertionsOk :: Bool

_contractState :: state

newtype Spec state a

The Spec monad is a state monad over the ModelState with reader and writer components to keep track of newly created symbolic tokens. It is used exclusively by the nextState function to model the effects of an action on the blockchain.

Constructors

Spec

Fields

unSpec :: WriterT SymCreationIndex (ReaderT (Var SymIndex) (State (ModelState state))) a

coerceSpec :: forall s s' a. Coercible s s' => Spec s a -> Spec s' a

contractState :: forall state state. Lens (ModelState state) (ModelState state) state state

Lens for the contract-specific part of the model state.

currentSlotL :: forall state. Lens' (ModelState state) SlotNo

balanceChangesL :: forall state. Lens' (ModelState state) (Map (AddressInEra Era) SymValue)

mintedL :: forall state. Lens' (ModelState state) SymValue

assertions :: forall state. Lens' (ModelState state) [(String, Bool)]

assertionsOk :: forall state. Lens' (ModelState state) Bool

symbolics :: forall state. Lens' (ModelState state) SymCollectionIndex

currentSlot :: Getter (ModelState state) SlotNo

Get the current slot.

Spec monad update functions: wait and waitUntil.

balanceChanges :: Getter (ModelState state) (Map (AddressInEra Era) SymValue)

Get the current wallet balance changes. These are delta balances, so they start out at zero and can be negative. The absolute balances used by the emulator can be set in the CheckOptions argument to propRunActionsWithOptions.

Spec monad update functions: withdraw, deposit, transfer.

balanceChange :: Ord (AddressInEra Era) => AddressInEra Era -> Getter (ModelState state) SymValue

Get the current balance change for a wallet. This is the delta balance, so it starts out at zero and can be negative. The absolute balance used by the emulator can be set in the CheckOptions argument to propRunActionsWithOptions.

Spec monad update functions: withdraw, deposit, transfer.

minted :: Getter (ModelState state) SymValue

Get the amount of tokens minted so far. This is used to compute lockedValue.

Spec monad update functions: mint and burn.

lockedValue :: ModelState s -> SymValue

How much value is currently locked by contracts. This computed by subtracting the wallet balances from the minted value.

modState :: forall state a. Setter' (ModelState state) a -> (a -> a) -> Spec state ()

Modify a field in the ModelState

class Monad m => GetModelState m where

Monads with read access to the model state: the Spec monad used in nextState, and the DL monad used to construct test scenarios.

Associated Types

type StateType m :: *

The contract state type of the monad. For both Spec and DL this is simply the state parameter of the respective monad.

Methods

getModelState :: m (ModelState (StateType m))

Get the current model state.

getContractState :: GetModelState m => m (StateType m)

Get the contract state part of the model state.

askModelState :: GetModelState m => (ModelState (StateType m) -> a) -> m a

Get a component of the model state.

askContractState :: GetModelState m => (StateType m -> a) -> m a

Get a component of the contract state.

viewModelState :: GetModelState m => Getting a (ModelState (StateType m)) a -> m a

Get a component of the model state using a lens.

viewContractState :: GetModelState m => Getting a (StateType m) a -> m a

Get a component of the contract state using a lens.

runSpec :: Spec state () -> Var SymIndex -> ModelState state -> ModelState state

symbolicsCreatedBy :: Spec state () -> Var SymIndex -> ModelState state -> SymCreationIndex

createSymbolic :: forall t state. HasSymbolicRep t => String -> Spec state (Symbolic t)

createToken :: String -> Spec state SymToken

Create a new symbolic token in nextState - must have a corresponding registerToken call in perform

createTxOut :: String -> Spec state SymTxOut

Create a new symbolic TxOut in nextState - must have a corresponding registerTxOut call in perform

createTxIn :: String -> Spec state SymTxIn

Create a new symbolic TxIn in nextState - must have a corresponding registerTxIn call in perform

mint :: SymValueLike v => v -> Spec state ()

Mint tokens. Minted tokens start out as lockedValue (i.e. owned by the contract) and can be transferred to wallets using deposit.

burn :: SymValueLike v => v -> Spec state ()

Burn tokens. Equivalent to mint . inv.

deposit :: SymValueLike v => AddressInEra Era -> v -> Spec state ()

Add tokens to the balanceChange of an address. The added tokens are subtracted from the lockedValue of tokens held by contracts.

withdraw :: SymValueLike v => AddressInEra Era -> v -> Spec state ()

Withdraw tokens from an address. The withdrawn tokens are added to the lockedValue of tokens held by contracts.

transfer

Arguments

:: SymValueLike v

=> AddressInEra Era
Transfer from this address

-> AddressInEra Era
to this address

-> v
this much value

-> Spec state ()

Transfer tokens between wallets, updating their balances.

assertSpec :: String -> Bool -> Spec state ()

Assert that a particular predicate holds at a point in the specification