Test.QuickCheck.ContractModel.Internal.Symbolics

Documentation

data SymToken

A symbolic token is a token that exists only during ContractModel generation time

Constructors

SymToken

Fields

symVar :: Var (Map String AssetId)

symVarIdx :: String

data SymValue

A symbolic value is a combination of a real value and a value associating symbolic tokens with an amount

Constructors

SymValue

Fields

symValMap :: Map SymToken Quantity

actualValPart :: Value

pPrintValue :: Value -> Doc

pPrintAssetId :: AssetId -> Doc

symIsZero :: SymValue -> Bool

Check if a symbolic value is zero

symLeq :: SymValue -> SymValue -> Bool

Check if one symbolic value is less than or equal to another

toValue :: (SymToken -> AssetId) -> SymValue -> Value

Using a semantics function for symbolic tokens, convert a SymValue to a Value

toSymVal :: (AssetId -> Maybe SymToken) -> Value -> SymValue

Invert a sym token mapping to turn a Value into a SymValue, useful for error reporting

inv :: SymValue -> SymValue

class SymValueLike v where

Methods

toSymValue :: v -> SymValue

class TokenLike t where

Methods

symAssetIdValueOf :: SymValue -> t -> Quantity

Get the value of a specific token in a SymValue

symAssetIdValue :: t -> Quantity -> SymValue

Convert a token and an amount to a SymValue