Test.QuickCheck.ContractModel

Documentation

class (Typeable state, Show state, HasActions state) => ContractModel state where

A ContractModel instance captures everything that is needed to generate and run tests of a contract or set of contracts. It specifies among other things

what operations are supported by the contract (Action),

when they are valid (precondition),

how to generate random actions (arbitraryAction),

how the operations affect the state (nextState), and

how to run the operations in the emulator (perform)

Minimal complete definition

arbitraryAction, initialState, nextState

Associated Types

data Action state

The type of actions that are supported by the contract. An action usually represents a single callEndpoint or a transfer of tokens, but it can be anything that can be interpreted in the EmulatorTrace monad.

Methods

arbitraryAction :: ModelState state -> Gen (Action state)

Given the current model state, provide a QuickCheck generator for a random next action. This is used in the Arbitrary instance for Actionss as well as by anyAction and anyActions.

actionName :: Action state -> String

The name of an Action, used to report statistics.

waitProbability :: ModelState state -> Double

The probability that we will generate a WaitUntil in a given state

arbitraryWaitInterval :: ModelState state -> Gen SlotNo

Control the distribution of how long WaitUntil waits

initialState :: state

The initial state, before any actions have been performed.

precondition :: ModelState state -> Action state -> Bool

The precondition function decides if a given action is valid in a given state. Typically actions generated by arbitraryAction will satisfy the precondition, but if they don't they will be discarded and another action will be generated. More importantly, the preconditions are used when shrinking (see shrinkAction) to ensure that shrunk test cases still make sense.

If an explicit action in a DL scenario violates the precondition an error is raised.

nextReactiveState :: SlotNo -> Spec state ()

nextReactiveState is run every time the model waits for a slot to be reached. This can be used to model reactive components of off-chain code.

nextState :: Action state -> Spec state ()

This is where the model logic is defined. Given an action, nextState specifies the effects running that action has on the model state. It runs in the Spec monad, which is a state monad over the ModelState.

shrinkAction :: ModelState state -> Action state -> [Action state]

When a test involving random sequences of actions fails, the framework tries to find a minimal failing test case by shrinking the original failure. Action sequences are shrunk by removing individual actions, or by replacing an action by one of the (simpler) actions returned by shrinkAction.

See shrink for more information on shrinking.

restricted :: Action state -> Bool

class (ContractModel state, IsRunnable m) => RunModel state m where

Minimal complete definition

perform

Methods

perform :: ModelState state -> Action state -> (forall t. HasSymbolicRep t => Symbolic t -> t) -> RunMonad m ()

Perform an Action in some state in the Monad m. This is the function that's used to exercise the actual stateful implementation, usually through various side-effects as permitted by m. It produces a value of type a, eg. some observable output from the Action that should later be kept in the environment through a `Var a` also passed to the nextState function.

The Lookup parameter provides an environment to lookup `Var a` instances from previous steps.

monitoring :: (ModelState state, ModelState state) -> Action state -> (forall t. HasSymbolicRep t => Symbolic t -> t) -> SymIndex -> Property -> Property

Allows the user to attach information to the Property at each step of the process. This function is given the full transition that's been executed, including the start and ending state, the Action, the current environment to Lookup and the value produced by perform while executing this step.

class (DefaultRealized m, HasChainIndex m, Monad m) => IsRunnable m where

Methods

awaitSlot :: SlotNo -> m ()

type DefaultRealized m = (Realized m SymIndex ~ SymIndex, Realized m () ~ ())

newtype RunMonad m a

Constructors

RunMonad

Fields

unRunMonad :: WriterT SymIndex m a

data Actions s

data Act s

Constructors

Bind

Fields

varOf :: Var SymIndex

actionOf :: Action s

NoBind

Fields

varOf :: Var SymIndex

actionOf :: Action s

ActWaitUntil (Var ()) SlotNo

ActObservation (Var ()) String (SymbolicSemantics -> ChainState -> Bool)

pattern Actions :: [Act s] -> Actions s

ContractAction :: Bool -> Action state -> Action (ModelState state) SymIndex

WaitUntil :: SlotNo -> Action (ModelState state) ()

stateAfter :: ContractModel state => Actions state -> ModelState state

runContractModel :: (ContractModel state, RunModel state m, HasChainIndex m) => Actions state -> PropertyM (RunMonad m) (ContractModelResult state)

liftRunMonad :: (forall a. m a -> n a) -> RunMonad m a -> RunMonad n a

contractState :: forall state state. Lens (ModelState state) (ModelState state) state state

Lens for the contract-specific part of the model state.

registerSymbolic :: (Monad m, HasSymbolicRep t) => String -> t -> RunMonad m ()

registerToken :: Monad m => String -> AssetId -> RunMonad m ()

registerTxOut :: Monad m => String -> TxOut CtxUTxO Era -> RunMonad m ()

registerTxIn :: Monad m => String -> TxIn -> RunMonad m ()

Chain index

class HasChainIndex m where

Methods

getChainIndex :: m ChainIndex

getChainState :: m ChainState

data ChainIndex

Constructors

ChainIndex

Fields

transactions :: [TxInState]

networkId :: NetworkId

data ChainState

Constructors

ChainState

Fields

slot :: SlotNo

utxo :: UTxO Era

data TxInState

Constructors

TxInState

Fields

tx :: Tx Era

chainState :: ChainState

accepted :: Bool

type Era = BabbageEra

The safe interface to Symbolic

data Symbolic t

type SymToken = Symbolic AssetId

A symbolic token is a token that is only available at runtime

type SymTxOut = Symbolic (TxOut CtxUTxO Era)

A SymTxOut is a `TxOut CtxUTxO Era` that is only available at runtime

type SymTxIn = Symbolic TxIn

A SymTxIn is a TxIn that is only available at runtime

data SymValue

A symbolic value is a combination of a real value and a value associating symbolic tokens with an amount

symIsZero :: SymValue -> Bool

Check if a symbolic value is zero

symLeq :: SymValue -> SymValue -> Bool

Check if one symbolic value is less than or equal to another

toValue :: (SymToken -> AssetId) -> SymValue -> Value

Using a semantics function for symbolic tokens, convert a SymValue to a Value

toSymVal :: (AssetId -> Maybe SymToken) -> Value -> SymValue

Invert a sym token mapping to turn a Value into a SymValue, useful for error reporting

inv :: SymValue -> SymValue

class SymValueLike v where

Methods

toSymValue :: v -> SymValue

class TokenLike t where

Methods

symAssetIdValueOf :: SymValue -> t -> Quantity

Get the value of a specific token in a SymValue

symAssetIdValue :: t -> Quantity -> SymValue

Convert a token and an amount to a SymValue

class HasSymbolics a where

Minimal complete definition

Nothing

Methods

getAllSymbolics :: a -> SymCollectionIndex

default getAllSymbolics :: (Generic a, GenericHasSymbolics (Rep a)) => a -> SymCollectionIndex

class HasSymbolicRep t

Minimal complete definition

symIndexL, symPrefix

Properties

data BalanceChangeOptions

Constructors

BalanceChangeOptions

Fields

observeScriptValue :: Bool

feeCalucation :: FeeCalculation

protocolParameters :: ProtocolParameters

addressPrettyPrinter :: AddressInEra Era -> String

assertBalanceChangesMatch :: BalanceChangeOptions -> ContractModelResult state -> Property

signerPaysFees :: FeeCalculation

asserts :: ModelState state -> Property

Dynamic logic

module Test.QuickCheck.ContractModel.DL

The safe interface to Spec

data ModelState state

The ModelState models the state of the blockchain. It contains,

the contract-specific state (contractState)

the current slot (currentSlot)

the wallet balances (balances)

the amount that has been minted (minted)

newtype Spec state a

The Spec monad is a state monad over the ModelState with reader and writer components to keep track of newly created symbolic tokens. It is used exclusively by the nextState function to model the effects of an action on the blockchain.

Constructors

Spec

Fields

unSpec :: WriterT SymCreationIndex (ReaderT (Var SymIndex) (State (ModelState state))) a

class Monad m => GetModelState m where

Monads with read access to the model state: the Spec monad used in nextState, and the DL monad used to construct test scenarios.

Associated Types

type StateType m :: *

The contract state type of the monad. For both Spec and DL this is simply the state parameter of the respective monad.

Methods

getModelState :: m (ModelState (StateType m))

Get the current model state.

runSpec :: Spec state () -> Var SymIndex -> ModelState state -> ModelState state

currentSlot :: Getter (ModelState state) SlotNo

Get the current slot.

Spec monad update functions: wait and waitUntil.

balanceChanges :: Getter (ModelState state) (Map (AddressInEra Era) SymValue)

Get the current wallet balance changes. These are delta balances, so they start out at zero and can be negative. The absolute balances used by the emulator can be set in the CheckOptions argument to propRunActionsWithOptions.

Spec monad update functions: withdraw, deposit, transfer.

balanceChange :: Ord (AddressInEra Era) => AddressInEra Era -> Getter (ModelState state) SymValue

Get the current balance change for a wallet. This is the delta balance, so it starts out at zero and can be negative. The absolute balance used by the emulator can be set in the CheckOptions argument to propRunActionsWithOptions.

Spec monad update functions: withdraw, deposit, transfer.

minted :: Getter (ModelState state) SymValue

Get the amount of tokens minted so far. This is used to compute lockedValue.

Spec monad update functions: mint and burn.

lockedValue :: ModelState s -> SymValue

How much value is currently locked by contracts. This computed by subtracting the wallet balances from the minted value.

getContractState :: GetModelState m => m (StateType m)

Get the contract state part of the model state.

askModelState :: GetModelState m => (ModelState (StateType m) -> a) -> m a

Get a component of the model state.

askContractState :: GetModelState m => (StateType m -> a) -> m a

Get a component of the contract state.

viewModelState :: GetModelState m => Getting a (ModelState (StateType m)) a -> m a

Get a component of the model state using a lens.

viewContractState :: GetModelState m => Getting a (StateType m) a -> m a

Get a component of the contract state using a lens.

createSymbolic :: forall t state. HasSymbolicRep t => String -> Spec state (Symbolic t)

createToken :: String -> Spec state SymToken

Create a new symbolic token in nextState - must have a corresponding registerToken call in perform

createTxOut :: String -> Spec state SymTxOut

Create a new symbolic TxOut in nextState - must have a corresponding registerTxOut call in perform

createTxIn :: String -> Spec state SymTxIn

Create a new symbolic TxIn in nextState - must have a corresponding registerTxIn call in perform

mint :: SymValueLike v => v -> Spec state ()

Mint tokens. Minted tokens start out as lockedValue (i.e. owned by the contract) and can be transferred to wallets using deposit.

burn :: SymValueLike v => v -> Spec state ()

Burn tokens. Equivalent to mint . inv.

deposit :: SymValueLike v => AddressInEra Era -> v -> Spec state ()

Add tokens to the balanceChange of an address. The added tokens are subtracted from the lockedValue of tokens held by contracts.

withdraw :: SymValueLike v => AddressInEra Era -> v -> Spec state ()

Withdraw tokens from an address. The withdrawn tokens are added to the lockedValue of tokens held by contracts.

transfer

Arguments

:: SymValueLike v

=> AddressInEra Era
Transfer from this address

-> AddressInEra Era
to this address

-> v
this much value

-> Spec state ()

Transfer tokens between wallets, updating their balances.

waitUntil :: ContractModel state => SlotNo -> Spec state ()

Wait until the given slot. Has no effect if currentSlot is greater than the given slot.

wait :: ContractModel state => Integer -> Spec state ()

Wait the given number of slots. Updates the currentSlot of the model state.

assertSpec :: String -> Bool -> Spec state ()

Assert that a particular predicate holds at a point in the specification

coerceSpec :: forall s s' a. Coercible s s' => Spec s a -> Spec s' a

Internals

fromStateModelActions :: Actions (ModelState s) -> Actions s