Test.QuickCheck.ContractModel.Internal

Documentation

class (ContractModel state, IsRunnable m) => RunModel state m where

Minimal complete definition

perform

Methods

perform :: ModelState state -> Action state -> (forall t. HasSymbolicRep t => Symbolic t -> t) -> RunMonad m ()

Perform an Action in some state in the Monad m. This is the function that's used to exercise the actual stateful implementation, usually through various side-effects as permitted by m. It produces a value of type a, eg. some observable output from the Action that should later be kept in the environment through a `Var a` also passed to the nextState function.

The Lookup parameter provides an environment to lookup `Var a` instances from previous steps.

monitoring :: (ModelState state, ModelState state) -> Action state -> (forall t. HasSymbolicRep t => Symbolic t -> t) -> SymIndex -> Property -> Property

Allows the user to attach information to the Property at each step of the process. This function is given the full transition that's been executed, including the start and ending state, the Action, the current environment to Lookup and the value produced by perform while executing this step.

newtype RunMonad m a

Constructors

RunMonad

Fields

unRunMonad :: WriterT SymIndex m a

liftRunMonad :: (forall a. m a -> n a) -> RunMonad m a -> RunMonad n a

registerSymbolic :: (Monad m, HasSymbolicRep t) => String -> t -> RunMonad m ()

registerToken :: Monad m => String -> AssetId -> RunMonad m ()

registerTxOut :: Monad m => String -> TxOut CtxUTxO Era -> RunMonad m ()

registerTxIn :: Monad m => String -> TxIn -> RunMonad m ()

withLocalSymbolics :: Monad m => RunMonad m () -> RunMonad m SymIndex

type DefaultRealized m = (Realized m SymIndex ~ SymIndex, Realized m () ~ ())

class (DefaultRealized m, HasChainIndex m, Monad m) => IsRunnable m where

Methods

awaitSlot :: SlotNo -> m ()

translateSymbolic :: (Var SymIndex -> SymIndex) -> SymbolicSemantics

data ContractModelResult state

Constructors

ContractModelResult

Fields

finalModelState :: ModelState state

symbolicTokens :: Map SymToken AssetId

finalChainIndex :: ChainIndex

runContractModel :: (ContractModel state, RunModel state m, HasChainIndex m) => Actions state -> PropertyM (RunMonad m) (ContractModelResult state)

data BalanceChangeOptions

Constructors

BalanceChangeOptions

Fields

observeScriptValue :: Bool

feeCalucation :: FeeCalculation

protocolParameters :: ProtocolParameters

addressPrettyPrinter :: AddressInEra Era -> String

assertBalanceChangesMatch :: BalanceChangeOptions -> ContractModelResult state -> Property

Orphan instances

(Monad m, HasChainIndex m) => HasChainIndex (PostconditionM m)

Instance details

Methods

getChainIndex :: PostconditionM m ChainIndex

getChainState :: PostconditionM m ChainState