Test.QuickCheck.ContractModel.Internal.Model

Documentation

class HasSymbolics a where

Minimal complete definition

Nothing

Methods

getAllSymbolics :: a -> SymCollectionIndex

default getAllSymbolics :: (Generic a, GenericHasSymbolics (Rep a)) => a -> SymCollectionIndex

newtype BaseType a

Constructors

BaseType a

class GenericHasSymbolics f where

Methods

genericGetAllSymbolics :: f k -> SymCollectionIndex

class (Typeable state, Show state, HasActions state) => ContractModel state where

A ContractModel instance captures everything that is needed to generate and run tests of a contract or set of contracts. It specifies among other things

what operations are supported by the contract (Action),

when they are valid (precondition),

how to generate random actions (arbitraryAction),

how the operations affect the state (nextState), and

how to run the operations in the emulator (perform)

Minimal complete definition

arbitraryAction, initialState, nextState

Associated Types

data Action state

The type of actions that are supported by the contract. An action usually represents a single callEndpoint or a transfer of tokens, but it can be anything that can be interpreted in the EmulatorTrace monad.

Methods

arbitraryAction :: ModelState state -> Gen (Action state)

Given the current model state, provide a QuickCheck generator for a random next action. This is used in the Arbitrary instance for Actionss as well as by anyAction and anyActions.

actionName :: Action state -> String

The name of an Action, used to report statistics.

waitProbability :: ModelState state -> Double

The probability that we will generate a WaitUntil in a given state

arbitraryWaitInterval :: ModelState state -> Gen SlotNo

Control the distribution of how long WaitUntil waits

initialState :: state

The initial state, before any actions have been performed.

precondition :: ModelState state -> Action state -> Bool

The precondition function decides if a given action is valid in a given state. Typically actions generated by arbitraryAction will satisfy the precondition, but if they don't they will be discarded and another action will be generated. More importantly, the preconditions are used when shrinking (see shrinkAction) to ensure that shrunk test cases still make sense.

If an explicit action in a DL scenario violates the precondition an error is raised.

nextReactiveState :: SlotNo -> Spec state ()

nextReactiveState is run every time the model waits for a slot to be reached. This can be used to model reactive components of off-chain code.

nextState :: Action state -> Spec state ()

This is where the model logic is defined. Given an action, nextState specifies the effects running that action has on the model state. It runs in the Spec monad, which is a state monad over the ModelState.

shrinkAction :: ModelState state -> Action state -> [Action state]

When a test involving random sequences of actions fails, the framework tries to find a minimal failing test case by shrinking the original failure. Action sequences are shrunk by removing individual actions, or by replacing an action by one of the (simpler) actions returned by shrinkAction.

See shrink for more information on shrinking.

restricted :: Action state -> Bool

createsSymbolics :: ContractModel state => ModelState state -> Action state -> Bool

Check if a given action creates new symbolic tokens in a given ModelState

wait :: ContractModel state => Integer -> Spec state ()

Wait the given number of slots. Updates the currentSlot of the model state.

waitUntil :: ContractModel state => SlotNo -> Spec state ()

Wait until the given slot. Has no effect if currentSlot is greater than the given slot.

contractAction :: ContractModel state => ModelState state -> Action state -> Action (ModelState state) SymIndex

data Actions s

Constructors

Actions_ [String] (Smart [Act s])

pattern ContractAction :: Bool -> Action state -> Action (ModelState state) SymIndex

pattern WaitUntil :: SlotNo -> Action (ModelState state) ()

pattern Observation :: String -> (SymbolicSemantics -> ChainState -> Bool) -> Action (ModelState state) ()

pattern Actions :: [Act s] -> Actions s

data Act s

Constructors

Bind

Fields

varOf :: Var SymIndex

actionOf :: Action s

NoBind

Fields

varOf :: Var SymIndex

actionOf :: Action s

ActWaitUntil (Var ()) SlotNo

ActObservation (Var ()) String (SymbolicSemantics -> ChainState -> Bool)

mapActions :: (Action s -> Action s') -> Actions s -> Actions s'

isBind :: Act s -> Bool

toStateModelActions :: ContractModel state => Actions state -> Actions (ModelState state)

fromStateModelActions :: Actions (ModelState s) -> Actions s

dummyModelState :: state -> ModelState state

stateAfter :: ContractModel state => Actions state -> ModelState state

annotatedStateAfter :: ContractModel state => Actions state -> Annotated (ModelState state)

asserts :: ModelState state -> Property

Orphan instances

HasVariables Quantity

Instance details

Methods

getAllVariables :: Quantity -> Set (Any Var)

HasVariables Value

Instance details

Methods

getAllVariables :: Value -> Set (Any Var)

ContractModel state => StateModel (ModelState state)

Instance details

Associated Types

data Action (ModelState state) a

Methods

actionName :: Action (ModelState state) a -> String

arbitraryAction :: VarContext -> ModelState state -> Gen (Any (Action (ModelState state)))

shrinkAction :: Typeable a => VarContext -> ModelState state -> Action (ModelState state) a -> [Any (Action (ModelState state))]

initialState :: ModelState state

nextState :: Typeable a => ModelState state -> Action (ModelState state) a -> Var a -> ModelState state

precondition :: ModelState state -> Action (ModelState state) a -> Bool

HasVariables (AddressInEra Era)

Instance details

Methods

getAllVariables :: AddressInEra Era -> Set (Any Var)

ContractModel state => Eq (Action (ModelState state) a)

Instance details

Methods

(==) :: Action (ModelState state) a -> Action (ModelState state) a -> Bool

(/=) :: Action (ModelState state) a -> Action (ModelState state) a -> Bool

ContractModel state => Show (Action (ModelState state) a)

Instance details

Methods

showsPrec :: Int -> Action (ModelState state) a -> ShowS

show :: Action (ModelState state) a -> String

showList :: [Action (ModelState state) a] -> ShowS

HasVariables (Action state) => HasVariables (Action (ModelState state) a)

Instance details

Methods

getAllVariables :: Action (ModelState state) a -> Set (Any Var)