Test.QuickCheck.ContractModel.Internal.Symbolics

Documentation

data SymIndexF f

Constructors

SymIndex

Fields

_tokens :: f AssetId

_utxos :: f (TxOut CtxUTxO Era)

_txIns :: f TxIn

utxos :: forall f. Lens' (SymIndexF f) (f (TxOut CtxUTxO Era))

txIns :: forall f. Lens' (SymIndexF f) (f TxIn)

tokens :: forall f. Lens' (SymIndexF f) (f AssetId)

class HasSymbolicRep t where

Methods

symIndexL :: Lens' (SymIndexF f) (f t)

symPrefix :: String

bmapConst :: FunctorB b => (forall a. f a -> c) -> b f -> Container b c

mappendSymIndexF :: forall f. (AllBF Semigroup f SymIndexF, Show (SymIndexF f)) => (forall a. f a -> Set String) -> SymIndexF f -> SymIndexF f -> SymIndexF f

type SymIndex = SymIndexF (Map String)

For an assumed variable, what's the mapping of String indices to underlying actual values. This is what is returned by a contract model action when it runs.

symIndex :: HasSymbolicRep t => String -> t -> SymIndex

type SymCreationIndex = SymIndexF (Const (Set String))

For a given action, what are the String indices used to construct symbolic variables when this action ran. NOTE: this purposefully does not include the variable because we might want to fake the variable with `Var 0` in some cases. See comment somewhere for why this is safe...

toCreationIndex :: SymIndex -> SymCreationIndex

createIndex :: forall t. HasSymbolicRep t => String -> SymCreationIndex

showCreateIndex :: SymCreationIndex -> String

type SymCollectionIndex = SymIndexF SymSet

What symbolic variables have been created in a given run of the Spec monad?

newtype SymSet t

Constructors

SymSet

Fields

unSymSet :: Set (Symbolic t)

symCollect :: HasSymbolicRep t => Symbolic t -> SymCollectionIndex

makeSymCollection :: SymCreationIndex -> Var SymIndex -> SymCollectionIndex

symCollectionSubset :: SymCollectionIndex -> SymCollectionIndex -> Bool

data Symbolic t

Constructors

Symbolic

Fields

symVar :: Var SymIndex

symVarIdx :: String

type SymbolicSemantics = forall t. HasSymbolicRep t => Symbolic t -> t

getSymbolics :: forall t. HasSymbolicRep t => SymCreationIndex -> Var SymIndex -> Set (Symbolic t)

lookupSymbolic :: HasSymbolicRep t => SymIndex -> Symbolic t -> Maybe t

type SymToken = Symbolic AssetId

A symbolic token is a token that is only available at runtime

type SymTxOut = Symbolic (TxOut CtxUTxO Era)

A SymTxOut is a `TxOut CtxUTxO Era` that is only available at runtime

type SymTxIn = Symbolic TxIn

A SymTxIn is a TxIn that is only available at runtime

data SymValue

A symbolic value is a combination of a real value and a value associating symbolic tokens with an amount

Constructors

SymValue

Fields

symValMap :: Map SymToken Quantity

actualValPart :: Value

pPrintValue :: Value -> Doc

pPrintAssetId :: AssetId -> Doc

symIsZero :: SymValue -> Bool

Check if a symbolic value is zero

symLeq :: SymValue -> SymValue -> Bool

Check if one symbolic value is less than or equal to another

toValue :: (SymToken -> AssetId) -> SymValue -> Value

Using a semantics function for symbolic tokens, convert a SymValue to a Value

toSymVal :: (AssetId -> Maybe SymToken) -> Value -> SymValue

Invert a sym token mapping to turn a Value into a SymValue, useful for error reporting

inv :: SymValue -> SymValue

class SymValueLike v where

Methods

toSymValue :: v -> SymValue

class TokenLike t where

Methods

symAssetIdValueOf :: SymValue -> t -> Quantity

Get the value of a specific token in a SymValue

symAssetIdValue :: t -> Quantity -> SymValue

Convert a token and an amount to a SymValue