Test.QuickCheck.ContractModel.ThreatModel.Cardano.Api

Documentation

addressOfTxOut :: TxOut ctx Era -> AddressAny

valueOfTxOut :: TxOut ctx Era -> Value

datumOfTxOut :: TxOut ctx Era -> TxOutDatum ctx Era

Get the datum from a transaction output.

redeemerOfTxIn :: Tx Era -> TxIn -> Maybe ScriptData

paymentCredentialToAddressAny :: PaymentCredential -> AddressAny

scriptAddressAny :: ScriptHash -> AddressAny

Construct a script address.

keyAddressAny :: Hash PaymentKey -> AddressAny

Construct a public key address.

isKeyAddressAny :: AddressAny -> Bool

Check if an address is a public key address.

recomputeScriptData :: Maybe Word64 -> (Word64 -> Word64) -> TxBodyScriptData Era -> TxBodyScriptData Era

emptyTxBodyScriptData :: TxBodyScriptData Era

addScriptData :: Word64 -> Data (ShelleyLedgerEra Era) -> (Data (ShelleyLedgerEra Era), ExUnits) -> TxBodyScriptData Era -> TxBodyScriptData Era

addDatum :: Data (ShelleyLedgerEra Era) -> TxBodyScriptData Era -> TxBodyScriptData Era

toCtxUTxODatum :: TxOutDatum CtxTx Era -> TxOutDatum CtxUTxO Era

txOutDatum :: ScriptData -> TxOutDatum CtxTx Era

Convert ScriptData to a Datum.

toScriptData :: ToData a => a -> ScriptData

Convert a Haskell value to ScriptData for use as a Redeemer or convert to a Datum with txOutDatum.

dummyTxId :: TxId

Used for new inputs.

makeTxOut :: AddressAny -> Value -> TxOutDatum CtxTx Era -> ReferenceScript Era -> TxOut CtxUTxO Era

txSigners :: Tx Era -> [Hash PaymentKey]

txInputs :: Tx Era -> [TxIn]

txOutputs :: Tx Era -> [TxOut CtxTx Era]

leqValue :: Value -> Value -> Bool

Check if a value is less or equal than another value.

projectAda :: Value -> Value

Keep only the Ada part of a value.

data ValidityReport

The result of validating a transaction. In case of failure, it includes a list of reasons.

Constructors

ValidityReport

Fields

valid :: Bool

errors :: [String]

validateTx :: ProtocolParameters -> Tx Era -> UTxO Era -> ValidityReport

restrictUTxO :: Tx Era -> UTxO Era -> UTxO Era

Keep only UTxOs mentioned in the given transaction.

convValidityInterval :: (TxValidityLowerBound era, TxValidityUpperBound era) -> ValidityInterval